+++ core/CMakeLists.txt
@@ -86,6 +86,8 @@ set(SUBSURFACE_CORE_LIB_SRCS
 	file.h
 	format.cpp
 	format.h
+	gas.c
+	gas.h
 	gas-model.c
 	gaspressures.c
 	gaspressures.h
